Split Action content cards in differents file for a better organization
