+++ apps/ledger-live-desktop/src/renderer/components/ContentCards/ActionCard/components.tsx
@@ -1,0 +1,49 @@
+import styled from "styled-components";
+
+const CardContainer = styled.div`
+  height: 64px;
+  padding: 0px 16px 0px 16px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 16px;
+`;
+
+const Header = styled.img`
+  height: 36px;
+  width: 36px;
+  border-radius: 1000px;
+  background-color: white;
+`;
+
+const Body = styled.div`
+  height: 40px;
+  flex-grow: 1;
+  flex-basis: 0;
+`;
+
+const Title = styled.div`
+  color: white;
+  font-weight: 600;
+  font-size: 14px;
+
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+`;
+
+const Subtitle = styled.div`
+  font-weight: 500;
+  font-size: 13px;
+
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+`;
+
+const Actions = styled.div`
+  display: flex;
+  gap: 12px;
+`;
+
+export { CardContainer, Header, Body, Title, Subtitle, Actions };
